[wip] Adding types for 'time-cache' library
There is a third-party library 'time-cache' that does not have types and is
causing all sorts of problems. Including some quite general types here to get
it working. If successful, I'll get it submitted to DefinitelyTyped so it's a
one-line install for the future and it won't need to be included in the repo.