fix!: encoder does not have enough information to deduce pubsub topic in autosharding
A decoder needs supplementary information such as the cluster id or the number of shards in the network to compute the pubsub topic.

This information is held by the Waku node (running on a specific network).
It was "working" because default value were applied.

This is the start to a larger clean-up in terms of sharding logic.