fix: metadata should not send pubsub topic
The of an edge "subscribed" to given pubsub topics is incorrect as an edge node may dynamically attempt to subscribe or send messages on various shards. Especially with autosharding where sharding is abstracted to the developer.

What really matter, is for an edge node to find service nodes that are supporting the right shards.

The only scenario that may matter is for relay, but gossipsub handles the pubsub topic peer filter and management.

Removing this, simplifies the code by removing the need of pubsub topic passing to medata.