Remove pubsub topic restriction and peer management optimization
There is a pre-optimization to the connection manager to only connect to
nodes "in the same shard".

This concept does not make sense for an edge node, an edge node does
belong to a "shard" as it does not run relay and hence subscribe to a
shard (and even a relay node may need to access other shards via light
protocols).

The concept of pubsub topic or shards at a node level are meaningless
for an edge node.

An edge node need to be able to fire up light push, store and filter
requests on any shard, depending on the application need.

Hence, no pubsub restriction should be applied and a node should
not "stop to connect" to nodes in the "wrong shard".

Finally, there may be an optimization to "curate a good number
of nodes per shards". But this seems pre-mature.

Why is this needed? because I am looking at improving the
autosharding vs static sharding logic. This simplifies follow-up
pull requests.

This commit moves away from the concept of pubsub topic, remove
the need of "pubsub topic" at node level and simplifies the code.